Front and back image pokemon getters. Shiny palettes

+++ src/main/java/me/hugmanrique/pokedata/utils/ImageUtils.java
@@ -11,18 +11,18 @@ import me.hugmanrique.pokedata.roms.ROM;
  * @since 01/06/2017
  */
 public class ImageUtils {
-    public static Palette getPalette(ROM rom, long pointer, ImageType type) {
-        int[] data = Lz77.decompress(rom, (int) pointer);
+    public static Palette getPalette(ROM rom, int pointer, ImageType type) {
+        int[] data = Lz77.decompress(rom, pointer);
 
         return new Palette(type, data);
     }
 
-    public static Palette getPalette(ROM rom, long pointer) {
+    public static Palette getPalette(ROM rom, int pointer) {
         return getPalette(rom, pointer, ImageType.C16);
     }
 
-    public static ROMImage getImage(ROM rom, long pointer, Palette palette, int width, int height) {
-        int[] data = Lz77.decompress(rom, (int) pointer);
+    public static ROMImage getImage(ROM rom, int pointer, Palette palette, int width, int height) {
+        int[] data = Lz77.decompress(rom, pointer);
 
         return new ROMImage(palette, data, width, height);
     }
